Update index.html
adding functions and code to keep track of absolute count, to track the image filenames across images folders. Added function o find the folder name based off of the absolute count. Added function to increment the fileName by one or reset it to 1 if >950. Modified the for loop to increment absolute count, increment fileName correctly and get the folderName every time it executes. (I can modify the getFolderName() in the future to where I don't have to run it each loop).

The loop should now only display 100 images from the appropriate folder depending on the querystring of the page.