Update queue_handler.js
changed the .createEntryButton() method to add an event listener for a click instead of setting the onclick attribute. Changed the passed functions in the corresponding children classes to be anonymous so hopefully I can rewrite functions to not rely on divId as a passed parameter. I might be able to rewrite my functions as methods of the entry class.