Update queue_handler_noList.js
changing how buttons are added so that event listeners are added after the fact in the child classes and not passed as args to the parent. Also trying to implement a function to get the index of an item in a given list since I can't .indexOf an element collection.