updating cookie functions
Updating the cookies to store url's instead of div ID's. Trying to implement a new function to get a main entry object from a url which I'm not sure will work.

These changes should hopefully preserve the correct entries in all cookies after the playlist log text file is updated, since the url's are unique and immutable and the div ID's change each time  the log is updated.