Removed the archive directories and added an entry  in the gitignore.
The new entry is to ignore a folder /archive/images which is where I would like the
acutal archive images to be uploaded. So that they are not included in the repo
and don't bloat the commits.

Removed the archive folders because their indexes weren't used and they weren't really
useful for anything.